Numbering Systems Converter
Project Overview:
This program takes a decimal input from the user and converts it to binary, octal, and in hexadecimal .You can run this program in a C environment or compiler to see the results. 
What we did? first of all we include our header files that will allows me to use specific functions, secondly, we set up some definitions so we can summarize the project and make it in short so it can be simple and applicable, finally we did the implementations of the conversion functions. They use the input decimal number and convert it to binary, octal and hexadecimal, respectively.